requirements.txt
The libraries used are:
1-pandas
2-numpy
3-matplotlib
4-tensorflow.keras
5-sklearn
6-google.colab